Rename LocalizationConfiguration to AdiutorLocalConfig and update uses
This change renames the LocalizationConfiguration class to
AdiutorLocalConfig to better reflect its purpose. All references in the
codebase are updated accordingly, including the SPECIALPAGES section in
extension.json. The module configurations' constants have been
renamed to be more concise and intuitive.

A redundant AdiutorSettings class file is also removed in favor of
housing the class in the Specials namespace, aligning with the standard
MediaWiki extension structure. Associated tests are updated to match
this change, ensuring that PHPUnit expectations correctly reference
Message class usage.

Change-Id: I73d4b18ec05c9613023efec7f6aa5c58caa19cb3