tests: Remove redundant WANObjectCache override
Since 2020 with change Ia60cb0bc18b23, the cache is now enabled in
PHPUnit context by default, so this override is identical to the
default. If it needs to override it, it should use setMainCache()
instead since WANObjectCache is just a wrapper around that, having
the two diverge seems confusing, but in this case no override is
needed.

Simplify some variable names, and add a FIXME/TODO as the test does
not seem to do anything useful right now.

Change-Id: I33b4dd79b4c1a06752c686283174d744eecb6827